% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgname.R
\docType{package}
\name{MMAPPR2-package}
\alias{MMAPPR2}
\alias{MMAPPR2-package}
\title{Mutation Mapping Analysis Pipeline for Pooled RNA-seq}
\description{
The main functionality of this package is described in the \code{\link{mamppr}}
function.
}
\details{
inheritSection mmappr Examples
}
\author{
\strong{Maintainer}: First Last \email{first.last@example.com}

}
