% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MMAPPR2.R
\docType{package}
\name{MMAPPR2}
\alias{MMAPPR2}
\title{Mutation Mapping Analysis Pipeline for Pooled RNA-seq}
\description{
The main functionality of this package is described in the \code{\link{mmappr}}
function.
}
