% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_methods.R
\docType{methods}
\name{MmapprData-getters}
\alias{MmapprData-getters}
\alias{param}
\alias{distance}
\alias{peaks}
\alias{candidates}
\alias{param,MmapprData-method}
\alias{distance,MmapprData-method}
\alias{peaks,MmapprData-method}
\alias{candidates,MmapprData-method}
\title{MmapprData Getters}
\usage{
\S4method{param}{MmapprData}(obj)

\S4method{distance}{MmapprData}(obj)

\S4method{peaks}{MmapprData}(obj)

\S4method{candidates}{MmapprData}(obj)
}
\arguments{
\item{obj}{Desired \code{\linkS4class{MmapprData}} object.}
}
\value{
Desired attribute.
}
\description{
Access slots of \code{\linkS4class{MmapprData}} object.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE) &
        Sys.which('vep') != '') {
    genDir <- gmapR::GmapGenomeDirectory(tempdir(), create=TRUE)
    
    param <- MmapprParam(refGenome = gmapR::GmapGenome("GRCz11", genDir),
                               wtFiles = MMAPPR2data::exampleWTbam(),
                               mutFiles = MMAPPR2data::exampleMutBam(),
                               species = "danio_rerio")

    md <- new('MmapprData', param = param)

    param(md)
    distance(md)
    peaks(md)
    candidates(md)
}
}
\seealso{
\code{\linkS4class{MmapprData}}
}
