% Manually synchronized with R/mmapprData.R.
\name{mmapprData}
\alias{mmapprData}
\title{MmapprData Constructor}
\usage{mmapprData(param)}
\arguments{\item{param}{A \code{MmapprParam} object.}}
\value{A new \code{MmapprData} object.}
\description{Creates the mutable-by-pipeline result container used to carry parameters, linkage data, peaks, and candidates through the MMAPPR2 stages.}
\seealso{\code{\link{mmappr}}, \code{\link{mmapprParam}}}
