% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_methods.R
\docType{methods}
\name{MmapprDataGetters}
\alias{MmapprDataGetters}
\alias{param}
\alias{distance}
\alias{peaks}
\alias{candidates}
\alias{param,MmapprData-method}
\alias{distance,MmapprData-method}
\alias{peaks,MmapprData-method}
\alias{candidates,MmapprData-method}
\title{MmapprData Getters}
\usage{
\S4method{param}{MmapprData}(obj)

\S4method{distance}{MmapprData}(obj)

\S4method{peaks}{MmapprData}(obj)

\S4method{candidates}{MmapprData}(obj)
}
\arguments{
\item{obj}{Desired \code{\linkS4class{MmapprData}} object.}
}
\description{
Access slots of \code{\linkS4class{MmapprData}} object.
}
\seealso{
\code{\linkS4class{MmapprData}}
}
