% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{postCalcDistMD}
\alias{postCalcDistMD}
\title{MmapprData object after \code{\link{calculateDistance}} step.}
\format{A \code{\linkS4class{MmapprData}} object.}
\source{
\url{https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3613585/}
}
\usage{
postCalcDistMD
}
\description{
Produced originally from BAM files of the zy13 mutant
and wild-type pools, aligned to the GRCz11 reference genome
using HISAT2, as available in the \link[MMAPPR2data]{MMAPPR2data}
package.
}
\keyword{datasets}
