% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bam.R
\name{readInFiles}
\alias{readInFiles}
\title{Read BAM files and generate Euclidean distance data}
\usage{
readInFiles(mmapprData)
}
\arguments{
\item{mmapprData}{}
}
\value{
A \linkS4class{MmapprData} object with the \code{distance}
  slot filled.
}
\description{
Read BAM files and generate Euclidean distance data
}
