% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_bam.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Read BAM files and generate Euclidean distance data}
\usage{
calculateDistance(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{distance}
  slot filled.
}
\description{
Initalizes the MMAPPR2 pipeline and precedes the \code{\link{loessFit}}
step.
}
\examples{
\dontrun{
md <- calculateDistance(md)
}
}
