% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Read BAM files and generate Euclidean distance data}
\usage{
calculateDistance(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{distance}
  slot filled.
}
\description{
First step in the MMAPPR2 pipeline. Precedes the \code{\link{loessFit}}
step.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    MMAPPR2:::.insertFakeVEPintoPath() ## Ignore this line
    
    mmapprParam <- MmapprParam(refGenome = gmapR::GmapGenome("GRCz11"),
                               wtFiles = MMAPPR2data::zy13wtBam(),
                               mutFiles = MMAPPR2data::zy13mutBam(),
                               species = "danio_rerio")

    md <- new('MmapprData', param = mmapprParam)
    postCalcDistMD <- calculateDistance(md)
})
}
