% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.R
\name{calculateDistance}
\alias{calculateDistance}
\title{Read BAM files and generate Euclidean distance data}
\usage{
calculateDistance(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{distance}
  slot filled.
}
\description{
First step in the MMAPPR2 pipeline. Precedes the \code{\link{loessFit}}
step.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly=TRUE)
        & all(Sys.which(c("samtools", "vep")) != "")) {
    mmappr_param <- MmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                                mutFiles = MMAPPR2data::exampleMutBam(),
                                refFasta = MMAPPR2data::goldenFasta(),
                                gtf = MMAPPR2data::gtf(),
                                outputFolder = tempOutputFolder())

    md <- new('MmapprData', param = mmappr_param)
    postCalcDistMD <- calculateDistance(md)
}

}
