% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{generateCandidates}
\alias{generateCandidates}
\title{Generate potential causative mutations and consequences in peak regions}
\usage{
generateCandidates(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{candidates}
  slot filled with a \code{\link[GenomicRanges]{GRanges}} object for each
  peak chromosome containing variants and predicted consequences from
  Ensembl's Variant Effect Predictor.
}
\description{
Follows the \code{\link{peakRefinement}} step and produces a
\code{\linkS4class{MmapprData}} object ready for
\code{\link{outputMmapprData}}.
}
\examples{
\dontrun{
postCandidatesMD <- generateCandidates(postPeakRefMD)
}
}
