% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{generateCandidates}
\alias{generateCandidates}
\title{Generate potential causative mutations in peak regions}
\usage{
generateCandidates(mmapprData)
}
\arguments{
\item{mmapprData}{}
}
\description{
Generate potential causative mutations in peak regions
}
