% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{generateCandidates}
\alias{generateCandidates}
\title{Generate potential causative mutations and consequences in peak regions}
\usage{
generateCandidates(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{candidates}
  slot filled with a \code{\link[GenomicRanges]{GRanges}} object for each
  peak chromosome containing variants and predicted consequences from
  Ensembl's Variant Effect Predictor.
}
\description{
Follows the \code{\link{peakRefinement}} step and produces a
\code{\linkS4class{MmapprData}} object ready for
\code{\link{outputMmapprData}}.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE) &
        Sys.which('vep') != '') {
    slc24a5genome <- gmapR::GmapGenome(MMAPPR2data::goldenFasta(),
                                       name = 'slc24a5',
                                       create = TRUE)
    # Specify parameters:
    mmapprParam <- MmapprParam(refGenome = slc24a5genome,
                               wtFiles = MMAPPR2data::exampleWTbam(),
                               mutFiles = MMAPPR2data::exampleMutBam(),
                               species = "danio_rerio",
                               outputFolder = tempdir())
}
\dontrun{
md <- new('MmapprData', param = mmapprParam)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)
postPrePeakMD <- prePeak(postLoessMD)
postPeakRefMD <- peakRefinement(postPrePeakMD)

postCandidatesMD <- generateCandidates(postPeakRefMD)
}
}
