% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/candidates.R
\name{generateCandidates}
\alias{generateCandidates}
\title{Generate candidate mutations and consequences in peak regions}
\usage{
Follows the \code{\link{peakRefinement}} step and produces a
\code{\linkS4class{md}} object ready for
\code{\link{outputmd}}.
}
\arguments{
\item{md}{The \code{\linkS4class{md}} object to be analyzed.}
}
\value{
A \code{\linkS4class{md}} object with the \code{candidates}
  slot filled with a \code{\link[GenomicRanges]{GRanges}} object for each
  peak chromosome containing variants and predicted consequences from
  Ensembl's Variant Effect Predictor.
}
\description{
Generate candidate mutations and consequences in peak regions
}
\examples{
if (requireNamespace('MMAPPR2data', quietly=TRUE)) {
    mmappr_param <- mmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                                mutFiles = MMAPPR2data::exampleMutBam(),
                                refFasta = MMAPPR2data::goldenFasta(),
                                gtf = MMAPPR2data::gtf(),
                                outputFolder = tempOutputFolder())
}

\dontrun{
md <- mmapprData(mmappr_param)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)
postPrePeakMD <- prePeak(postLoessMD)
postPeakRefMD <- peakRefinement(postPrePeakMD)

postCandidatesMD <- generateCandidates(postPeakRefMD)
}

}
