% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicc_loess.R
\name{loessFit}
\alias{loessFit}
\title{Perform optimized Loess regression for each chromosome}
\usage{
loessFit(mmapprData)
}
\arguments{
\item{mmapprData}{}
}
\value{
A \linkS4class{MmapprData} object with the \code{$loess} element
  of the \code{distance} slot list filled.
}
\description{
Perform optimized Loess regression for each chromosome
}
