% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aicc_loess.R
\name{loessFit}
\alias{loessFit}
\title{Perform optimized Loess regression for each chromosome}
\usage{
loessFit(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{$loess} element
  of the \code{distance} slot list filled.
}
\description{
Called after the \code{\link{calculateDistance}} step and before
\code{\link{prePeak}}.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE) &
        Sys.which('vep') != '') {
    slc24a5genome <- gmapR::GmapGenome(MMAPPR2data::goldenFasta(),
                                       name = 'slc24a5',
                                       create = TRUE)
    # Specify parameters:
    mmapprParam <- MmapprParam(refGenome = slc24a5genome,
                               wtFiles = MMAPPR2data::exampleWTbam(),
                               mutFiles = MMAPPR2data::exampleMutBam(),
                               species = "danio_rerio",
                               outputFolder = tempOutputFolder())
}
\dontrun{
md <- new('MmapprData', param = mmapprParam)
postCalcDistMD <- calculateDistance(md)

postLoessMD <- loessFit(postCalcDistMD)
}
}
