% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loess.R
\name{loessFit}
\alias{loessFit}
\title{Perform optimized Loess regression for each chromosome}
\usage{
loessFit(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{$loess} 
  element of the \code{distance} slot list filled.
}
\description{
Called after the \code{\link{calculateDistance}} step and before
\code{\link{prePeak}}.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    mmappr_param <- mmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                                mutFiles = MMAPPR2data::exampleMutBam(),
                                refFasta = MMAPPR2data::goldenFasta(),
                                gtf = MMAPPR2data::exampleGTF(),
                                outputFolder = tempOutputFolder())
}

\dontrun{
md <- mmapprData(mmappr_param)
postCalcDistMD <- calculateDistance(md)

postLoessMD <- loessFit(postCalcDistMD)
}

}
