% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mmappr}
\alias{mmappr}
\title{Mutation Mapping Analysis Pipeline for Pooled RNA-Seq}
\usage{
mmappr(mmapprParam, showDebug = FALSE)
}
\arguments{
\item{mmapprParam}{A \linkS4class{MmapprParam} object containing desired parameters}

\item{showDebug}{If set to \code{TRUE}, intermediate dataframe sizes are
displayed during file reading step}
}
\value{
A \linkS4class{MmapprData} object containing results
}
\description{
This describes what we're doing
}
\examples{
\dontrun{
vepFlags <- ensemblVEP::VEPFlags(
                 flags=list(species = "danio_rerio_merged", 
                            format = "vcf",
                            database = TRUE))
mmapprParam <- MmapprParam(GmapGenome("danio_rerio_10"),
                           "wild_type.sorted.bam",
                           "mutant.sorted.bam",
                           vepFlags)
mmapprData <- mmappr(mmapprParam)
}
}
