% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mmappr}
\alias{mmappr}
\title{Mutation Mapping Analysis Pipeline for Pooled RNA-Seq}
\usage{
mmappr(mmapprParam)
}
\arguments{
\item{mmapprParam}{A \code{\linkS4class{MmapprParam}} object containing
desired parameters.}
}
\value{
A \code{\linkS4class{MmapprData}} object containing results
  as well as intermediate data.
}
\description{
MMAPPR2 is designed to map the causative mutation in a forward genetics
screen. It analyzes aligned sequence files, calculates the per-base
Euclidean distance between the mutant and wild-type pools, performs
a Loess regression on that distance, and generates candidate variants
in regions of peak distance.
}
\examples{
\dontrun{
mmapprParam <- MmapprParam(refGenome = GmapGenome("GRCz11"),
                           wtFiles = "wild_type.sorted.bam",
                           mutFiles = "mutant.sorted.bam",
                           species = "danio_rerio")
mmapprData <- mmappr(mmapprParam)

### Alternately, you can navigate the pipeline step by step.
### This may be helpful for debugging.
md <- new('MmapprData', param = mmapprParam)
md <- calculateDistance(md)
md <- loessFit(md)
md <- prePeak(md)
md <- peakRefinement(md)
md <- generateCandidates(md)
md <- outputMmapprData(md)
}

}
\seealso{
\code{\link{calculateDistance}}, \code{\link{loessFit}},
  \code{\link{prePeak}}, \code{\link{peakRefinement}},
  \code{\link{generateCandidates}}, \code{\link{outputMmapprData}}
}
