% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{mmappr}
\alias{mmappr}
\title{Mutation Mapping Analysis Pipeline for Pooled RNA-Seq}
\usage{
mmappr(mmapprParam)
}
\arguments{
\item{mmapprParam}{A \code{\linkS4class{MmapprParam}} object containing
desired parameters.}
}
\value{
A \code{\linkS4class{MmapprData}} object containing results
  and/or intermediate data.
}
\description{
MMAPPR2 is designed to map the causative mutation in a forward genetics
screen. It analyzes aligned sequence files, calculates the per-base
Euclidean distance between the mutant and wild-type pools, performs
a Loess regression on that distance, and generates candidate variants
in regions of peak distance.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    ## Ignore these lines:
    MMAPPR2:::.insertFakeVEPintoPath()
    genDir <- gmapR::GmapGenomeDirectory('DEFAULT', create=TRUE)

    # Specify parameters:
    mmapprParam <- MmapprParam(refGenome = gmapR::GmapGenome("GRCz11", genDir),
                               wtFiles = MMAPPR2data::zy13wtBam(),
                               mutFiles = MMAPPR2data::zy13mutBam(),
                               species = "danio_rerio")
                               
    # Run pipeline:
    mmapprData <- mmappr(mmapprParam)

}
=======

    # Specify parameters:
    mmapprParam <- MmapprParam(refGenome = gmapR::GmapGenome("GRCz11", genDir),
                               wtFiles = MMAPPR2data::zy13wtBam(),
                               mutFiles = MMAPPR2data::zy13mutBam(),
                               species = "danio_rerio")
                               
    # Run pipeline:
    mmapprData <- mmappr(mmapprParam)

}
\dontrun{
### Alternately, you can navigate the pipeline step by step.
### This may be helpful for debugging.
md <- new('MmapprData', param = mmapprParam)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)
postPrePeakMD <- prePeak(postLoessMD)
postPeakRefMD <- peakRefinement(postPrePeakMD)
postCandidatesMD <- generateCandidates(postPeakRefMD)
outputMmapprData(postCandidatesMD)
} 
>>>>>>> origin

}
\seealso{
\code{\link{calculateDistance}}, \code{\link{loessFit}},
  \code{\link{prePeak}}, \code{\link{peakRefinement}},
  \code{\link{generateCandidates}}, \code{\link{outputMmapprData}}
}
