% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R
\docType{class}
\name{MmapprData-class}
\alias{MmapprData-class}
\title{MmapprData object}
\description{
Stores data from each step of the MMAPPR pipeline.
}
\section{Slots}{

\describe{
\item{\code{param}}{MmapprParam. TODO}

\item{\code{distance}}{list.}

\item{\code{peaks}}{list.}

\item{\code{candidates}}{list.}
}}

