% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmapprData.R
\name{MmapprData-Class}
\alias{MmapprData-Class}
\alias{MmapprData}
\title{MmapprData Class}
\usage{
Stores data from each step of the MMAPPR2 pipeline.
}
\description{
MmapprData Class
}
\section{Slots}{

\describe{
\item{\code{param}}{\code{\linkS4class{MmapprParam}} object storing parameters
used in analysis.}

\item{\code{snpDistance}}{List containing raw counts and Euclidean distance data for
each chromosome. After \code{\link{calculateDistance}}, chromosomes with
sufficient data should have \code{$wtCounts},
\code{$mutCounts}, and \code{$distanceDf} populated. After
\code{\link{loessFit}}, the \code{$distanceDf} element for each chromosome
list is replaced with a \code{$loess} element.}

\item{\code{peaks}}{List of chromosomes containing peak regions. Initialized after
\code{\link{prePeak}} and populated with density function after
\code{\link{peakRefinement}}.}

\item{\code{candidates}}{List containing \code{\link[GenomicRanges]{GRanges}} objects
with snps, VEP predicted effects, and differential expression data
for each peak, all resulting from \code{\link{generateCandidates}} 
function.}
}}

\seealso{
\code{\link{mmappr}}, \link{MmapprData-getters}
}
