% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_methods.R
\name{MmapprData-functions}
\alias{MmapprData-functions}
\alias{param}
\alias{distance}
\alias{peaks}
\alias{candidates}
\alias{param,MmapprData-method}
\alias{snpDistance,MmapprData-method}
\alias{peaks,MmapprData-method}
\alias{candidates,MmapprData-method}
\alias{param<-,MmapprData-method}
\title{MmapprData Getters and Setters}
\usage{
\S4method{param}{MmapprData}(obj)

\S4method{snpDistance}{MmapprData}(obj)

\S4method{peaks}{MmapprData}(obj)

\S4method{candidates}{MmapprData}(obj)

\S4method{param}{MmapprData}(obj) <- value
}
\arguments{
\item{obj}{Desired \code{\linkS4class{MmapprData}} object.}
}
\value{
Desired attribute.
}
\description{
Access slots of \code{\linkS4class{MmapprData}} object. The only slot with 
a setter is "param," as the others must be generated using one of the 
functions in the pipeline.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)
        & all(Sys.which(c("samtools", "vep")) != "")) {
    mmappr_param <- MmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                                mutFiles = MMAPPR2data::exampleMutBam(),
                                refFasta = MMAPPR2data::goldenFasta(),
                                gtf = MMAPPR2data::gtf(),
                                outputFolder = tempOutputFolder())

    md <- new('MmapprData', param = mmappr_param)

    param(md)
    snpDistance(md)
    peaks(md)
    candidates(md)
}
}
\seealso{
\code{\linkS4class{MmapprData}}
}
