% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmapprData.R
\name{mmapprData}
\alias{mmapprData}
\title{MmapprData Constructor}
\usage{
Creates a \code{MmapprData} class Object. This object stores data from each
step of the MMAPPR2 pipeline.
}
\arguments{
\item{param}{\code{\linkS4class{MmapprParam}} object storing parameters
used in analysis.}
}
\value{
A \code{MmapprData} object.
}
\description{
MmapprData Constructor
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
mmappr_param <- mmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                            mutFiles = MMAPPR2data::exampleMutBam(),
                            refFasta = MMAPPR2data::goldenFasta(),
                            gtf = MMAPPR2data::gtf(),
                            outputFolder = tempOutputFolder())
}

mmapprData(mmappr_param)

}
\seealso{
\code{\linkS4class{MmapprData}} \code{\linkS4class{MmapprParam}}
}
