% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s4_methods.R
\docType{methods}
\name{MmapprParam-functions}
\alias{MmapprParam-functions}
\alias{refFasta}
\alias{refFasta<-}
\alias{wtFiles}
\alias{wtFiles<-}
\alias{mutFiles}
\alias{mutFiles<-}
\alias{species}
\alias{species<-}
\alias{vepFlags}
\alias{vepFlags<-}
\alias{refGenome}
\alias{refGenome<-}
\alias{homozygoteCutoff}
\alias{homozygoteCutoff<-}
\alias{distancePower}
\alias{distancePower<-}
\alias{peakIntervalWidth}
\alias{peakIntervalWidth<-}
\alias{minDepth}
\alias{minDepth<-}
\alias{minBaseQuality}
\alias{minBaseQuality<-}
\alias{minMapQuality}
\alias{minMapQuality<-}
\alias{loessOptResolution}
\alias{loessOptResolution<-}
\alias{loessOptCutFactor}
\alias{loessOptCutFactor<-}
\alias{naCutoff}
\alias{naCutoff<-}
\alias{outputFolder}
\alias{outputFolder<-}
\alias{fileAggregation}
\alias{fileAggregation<-}
\alias{refFasta,MmapprParam-method}
\alias{wtFiles,MmapprParam-method}
\alias{mutFiles,MmapprParam-method}
\alias{species,MmapprParam-method}
\alias{vepFlags,MmapprParam-method}
\alias{refGenome,MmapprParam-method}
\alias{homozygoteCutoff,MmapprParam-method}
\alias{distancePower,MmapprParam-method}
\alias{peakIntervalWidth,MmapprParam-method}
\alias{minDepth,MmapprParam-method}
\alias{minBaseQuality,MmapprParam-method}
\alias{minMapQuality,MmapprParam-method}
\alias{loessOptResolution,MmapprParam-method}
\alias{loessOptCutFactor,MmapprParam-method}
\alias{naCutoff,MmapprParam-method}
\alias{outputFolder,MmapprParam-method}
\alias{fileAggregation,MmapprParam-method}
\alias{refFasta<-,MmapprParam-method}
\alias{wtFiles<-,MmapprParam-method}
\alias{mutFiles<-,MmapprParam-method}
\alias{vepFlags<-,MmapprParam-method}
\alias{refGenome<-,MmapprParam-method}
\alias{species<-,MmapprParam-method}
\alias{homozygoteCutoff<-,MmapprParam-method}
\alias{distancePower<-,MmapprParam-method}
\alias{peakIntervalWidth<-,MmapprParam-method}
\alias{minDepth<-,MmapprParam-method}
\alias{minBaseQuality<-,MmapprParam-method}
\alias{loessOptResolution<-,MmapprParam-method}
\alias{loessOptCutFactor<-,MmapprParam-method}
\alias{naCutoff<-,MmapprParam-method}
\alias{outputFolder<-,MmapprParam-method}
\alias{minMapQuality<-,MmapprParam-method}
\alias{fileAggregation<-,MmapprParam-method}
\title{MmapprParam Getters and Setters}
\usage{
\S4method{refFasta}{MmapprParam}(obj)

\S4method{wtFiles}{MmapprParam}(obj)

\S4method{mutFiles}{MmapprParam}(obj)

\S4method{species}{MmapprParam}(obj)

\S4method{vepFlags}{MmapprParam}(obj)

\S4method{refGenome}{MmapprParam}(obj)

\S4method{homozygoteCutoff}{MmapprParam}(obj)

\S4method{distancePower}{MmapprParam}(obj)

\S4method{peakIntervalWidth}{MmapprParam}(obj)

\S4method{minDepth}{MmapprParam}(obj)

\S4method{minBaseQuality}{MmapprParam}(obj)

\S4method{minMapQuality}{MmapprParam}(obj)

\S4method{loessOptResolution}{MmapprParam}(obj)

\S4method{loessOptCutFactor}{MmapprParam}(obj)

\S4method{naCutoff}{MmapprParam}(obj)

\S4method{outputFolder}{MmapprParam}(obj)

\S4method{fileAggregation}{MmapprParam}(obj)

\S4method{refFasta}{MmapprParam}(obj) <- value

\S4method{wtFiles}{MmapprParam}(obj) <- value

\S4method{mutFiles}{MmapprParam}(obj) <- value

\S4method{vepFlags}{MmapprParam}(obj) <- value

\S4method{refGenome}{MmapprParam}(obj) <- value

\S4method{species}{MmapprParam}(obj) <- value

\S4method{homozygoteCutoff}{MmapprParam}(obj) <- value

\S4method{distancePower}{MmapprParam}(obj) <- value

\S4method{peakIntervalWidth}{MmapprParam}(obj) <- value

\S4method{minDepth}{MmapprParam}(obj) <- value

\S4method{minBaseQuality}{MmapprParam}(obj) <- value

\S4method{loessOptResolution}{MmapprParam}(obj) <- value

\S4method{loessOptCutFactor}{MmapprParam}(obj) <- value

\S4method{naCutoff}{MmapprParam}(obj) <- value

\S4method{outputFolder}{MmapprParam}(obj) <- value

\S4method{minMapQuality}{MmapprParam}(obj) <- value

\S4method{fileAggregation}{MmapprParam}(obj) <- value
}
\arguments{
\item{obj}{Desired \code{\link{MmapprParam}} object.}

\item{value}{Value to replace desired attribute.}
}
\value{
The desired \code{\link{MmapprParam}} attribute.
}
\description{
Access and assign slots of \code{\link{MmapprParam}} object.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly=TRUE)
        & Sys.which('vep') != '') {
    mmappr_param <- MmapprParam(refFasta = MMAPPR2data::goldenFasta(),
                         wtFiles = MMAPPR2data::exampleWTbam(),
                         mutFiles = MMAPPR2data::exampleMutBam(),
                         species = "danio_rerio")

    outputFolder(mmappr_param) <- 'mmappr2_test_1'
    minBaseQuality(mmappr_param) <- 25
    vepFlags(mmappr_param)
}
}
\seealso{
\code{\link{MmapprParam}}
}
