% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{MmapprParam-functions}
\alias{MmapprParam-functions}
\alias{wtFiles}
\alias{wtFiles<-}
\alias{mutFiles}
\alias{mutFiles<-}
\alias{refFasta}
\alias{refFasta<-}
\alias{gtf}
\alias{gtf<-}
\alias{outputFolder}
\alias{outputFolder<-}
\alias{includeScaffolds}
\alias{includeScaffolds<-}
\alias{minDepth}
\alias{minDepth<-}
\alias{homozygoteCutoff}
\alias{homozygoteCutoff<-}
\alias{minBaseQuality}
\alias{minBaseQuality<-}
\alias{minMapQuality}
\alias{minMapQuality<-}
\alias{fileAggregation}
\alias{fileAggregation<-}
\alias{distancePower}
\alias{distancePower<-}
\alias{peakIntervalWidth}
\alias{peakIntervalWidth<-}
\alias{loessOptResolution}
\alias{loessOptResolution<-}
\alias{loessOptCutFactor}
\alias{loessOptCutFactor<-}
\alias{wtFiles,MmapprParam-method}
\alias{mutFiles,MmapprParam-method}
\alias{refFasta,MmapprParam-method}
\alias{gtf,MmapprParam-method}
\alias{outputFolder,MmapprParam-method}
\alias{includeScaffolds,MmapprParam-method}
\alias{minDepth,MmapprParam-method}
\alias{homozygoteCutoff,MmapprParam-method}
\alias{minBaseQuality,MmapprParam-method}
\alias{minMapQuality,MmapprParam-method}
\alias{fileAggregation,MmapprParam-method}
\alias{distancePower,MmapprParam-method}
\alias{peakIntervalWidth,MmapprParam-method}
\alias{loessOptResolution,MmapprParam-method}
\alias{loessOptCutFactor,MmapprParam-method}
\alias{wtFiles<-,MmapprParam-method}
\alias{mutFiles<-,MmapprParam-method}
\alias{refFasta<-,MmapprParam-method}
\alias{gtf<-,MmapprParam-method}
\alias{outputFolder<-,MmapprParam-method}
\alias{includeScaffolds<-,MmapprParam-method}
\alias{minDepth<-,MmapprParam-method}
\alias{homozygoteCutoff<-,MmapprParam-method}
\alias{minBaseQuality<-,MmapprParam-method}
\alias{minMapQuality<-,MmapprParam-method}
\alias{fileAggregation<-,MmapprParam-method}
\alias{distancePower<-,MmapprParam-method}
\alias{peakIntervalWidth<-,MmapprParam-method}
\alias{loessOptResolution<-,MmapprParam-method}
\alias{loessOptCutFactor<-,MmapprParam-method}
\title{MmapprParam Getters and Setters}
\usage{
\S4method{wtFiles}{MmapprParam}(obj)

\S4method{mutFiles}{MmapprParam}(obj)

\S4method{refFasta}{MmapprParam}(obj)

\S4method{gtf}{MmapprParam}(obj)

\S4method{outputFolder}{MmapprParam}(obj)

\S4method{includeScaffolds}{MmapprParam}(obj)

\S4method{minDepth}{MmapprParam}(obj)

\S4method{homozygoteCutoff}{MmapprParam}(obj)

\S4method{minBaseQuality}{MmapprParam}(obj)

\S4method{minMapQuality}{MmapprParam}(obj)

\S4method{fileAggregation}{MmapprParam}(obj)

\S4method{distancePower}{MmapprParam}(obj)

\S4method{peakIntervalWidth}{MmapprParam}(obj)

\S4method{loessOptResolution}{MmapprParam}(obj)

\S4method{loessOptCutFactor}{MmapprParam}(obj)

\S4method{wtFiles}{MmapprParam}(obj) <- value

\S4method{mutFiles}{MmapprParam}(obj) <- value

\S4method{refFasta}{MmapprParam}(obj) <- value

\S4method{gtf}{MmapprParam}(obj) <- value

\S4method{outputFolder}{MmapprParam}(obj) <- value

\S4method{includeScaffolds}{MmapprParam}(obj) <- value

\S4method{minDepth}{MmapprParam}(obj) <- value

\S4method{homozygoteCutoff}{MmapprParam}(obj) <- value

\S4method{minBaseQuality}{MmapprParam}(obj) <- value

\S4method{minMapQuality}{MmapprParam}(obj) <- value

\S4method{fileAggregation}{MmapprParam}(obj) <- value

\S4method{distancePower}{MmapprParam}(obj) <- value

\S4method{peakIntervalWidth}{MmapprParam}(obj) <- value

\S4method{loessOptResolution}{MmapprParam}(obj) <- value

\S4method{loessOptCutFactor}{MmapprParam}(obj) <- value
}
\arguments{
\item{obj}{Desired \code{\link{MmapprParam}} object.}

\item{value}{Value to replace desired attribute.}
}
\value{
The desired \code{\link{MmapprParam}} attribute.
}
\description{
Access and assign slots of \code{\link{MmapprParam}} object.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    mmappr_param <- mmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                         mutFiles = MMAPPR2data::exampleMutBam(),
                         refFasta = MMAPPR2data::goldenFasta(),
                         gtf = MMAPPR2data::gtf())

    outputFolder(mmappr_param) <- 'mmappr2_test_1'
    minBaseQuality(mmappr_param) <- 25
    vepFlags(mmappr_param)
}
}
\seealso{
\code{\link{MmapprParam}}
}
