% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_classes.R, R/s4_methods.R
\docType{class}
\name{MmapprParam-class}
\alias{MmapprParam-class}
\alias{MmapprParam}
\title{MmapprParam Class}
\usage{
MmapprParam(refGenome, wtFiles, mutFiles, vepFlags, outputFolder = "DEFAULT",
  distancePower = 4, peakIntervalWidth = 0.95, minDepth = 10,
  homozygoteCutoff = 0.95, minBaseQuality = 20, minMapQuality = 30,
  loessOptResolution = 0.001, loessOptCutFactor = 0.1, naCutoff = 0,
  fileAggregation = c("sum", "mean"))
}
\arguments{
\item{refGenome}{}

\item{wtFiles}{}

\item{mutFiles}{}

\item{vepFlags}{}

\item{outputFolder}{}

\item{distancePower}{}

\item{peakIntervalWidth}{}

\item{minDepth}{}

\item{homozygoteCutoff}{}

\item{minBaseQuality}{}

\item{minMapQuality}{}

\item{loessOptResolution}{}

\item{loessOptCutFactor}{}

\item{naCutoff}{}

\item{fileAggregation}{A length-one character vector determining strategy
for aggregating base calls when multiple wild-type or multiple mutant
files are provided.
When 'sum', average base call proportions are computed using
the read counts from each file, effectively weighting files
with higher counts at a given position. When equal to 'mean', the
base call proportions as well as read depths, rather than the absolute count,
are averaged across files, which is useful when you want to weight each
replicate evenly without regards to differing depth.}
}
\value{
A MmapprParam object
}
\description{
\code{MmapprParam} stores parameters running \code{\link{mmappr}}

\strong{Constructor}
}
\section{Slots}{

\describe{
\item{\code{refGenome}}{GmapGenome.}

\item{\code{wtFiles}}{BamFileList.}

\item{\code{mutFiles}}{BamFileList.}

\item{\code{vepFlags}}{VEPFlags.}

\item{\code{distancePower}}{numeric.}

\item{\code{peakIntervalWidth}}{numeric.}

\item{\code{minDepth}}{numeric.}

\item{\code{homozygoteCutoff}}{numeric.}

\item{\code{minBaseQuality}}{numeric.}

\item{\code{minMapQuality}}{numeric.}

\item{\code{loessOptResolution}}{numeric.}

\item{\code{loessOptCutFactor}}{numeric.}

\item{\code{naCutoff}}{numeric.}

\item{\code{outputFolder}}{character.}

\item{\code{fileAggregation}}{character.}
}}

\examples{
\dontrun{
mp <- MmapprParam(gmapR::GmapGenome('GRCz11'), 'wt.bam', 'mut.bam', VEPFlags())
}
}
