% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param.R
\name{mmapprParam}
\alias{mmapprParam}
\title{MmapprParam Constructor}
\usage{
Creates a new instance of a \code{\linkS4class{MmapprParam}} class object.
}
\arguments{
\item{wtFiles}{Character vector,
\code{\link[Rsamtools]{BamFile}}, or
\code{\link[Rsamtools]{BamFileList}} containing
BAM files for the wild-type pool to be analyzed.}

\item{mutFiles}{Character vector,
\code{\link[Rsamtools]{BamFile}}, or
\code{\link[Rsamtools]{BamFileList}} containing
BAM files for the mutant pool to be analyzed.}

\item{refFasta}{The path to the fasta file genome.}

\item{gtf}{The path to a gtf-formatted annotation file.}

\item{outputFolder}{Length-one character vector specifying where to save
output, including a \code{\linkS4class{MmapprData}} stored as
\code{mmappr_data.RDS}, \code{mmappr2.log}, a \code{.tsv} file
for each peak chromosome containing candidate mutations, and PDF plots
of both the entire genome and peak chromosomes. Defaults to an
automatically generated \code{mmappr2_<timestamp>}.}

\item{includeScaffolds}{Logical indicating whether non-standard chromosomes
should be included. False (default) trims the seqnames of the reference to 
include only standard chromosome names based on UCSC or Ensembl 
conventions. In all cases, the mitochondrial chromosome is removed.}

\item{minDepth}{Length-one integer vector determining minimum depth
required for a position to
be considered in the analysis. Defaults to 20.}

\item{homozygoteCutoff}{Length-one numeric vector between \code{0} and
\code{1} specifying threshold for throwing out base pairs on account
of homozygosity. Positions with high major allele frequency in the
wild-type pool are unlikely to exhibit polymorphism and are thus thrown
out when they exceed this cutoff. Defaults to \code{0.95}.}

\item{minBaseQuality}{Length-one numeric vector indicating minimum base
call quality to consider in analysis. Read positions with qualities
below this score will be thrown out. Defaults to 20.}

\item{minMapQuality}{Length-one numeric vector indicating minimum read
mapping quality to consider in analysis. Reads with qualities below
this score will be thrown out. Defaults to 30.}

\item{fileAggregation}{A length-one character vector determining strategy
for aggregating base calls when multiple wild-type or multiple mutant
files are provided.
When 'sum', average base call proportions are computed using
the read counts from each file, effectively weighting files
with higher counts at a given position. When equal to 'mean', the
base call proportions as well as read depths, rather than the absolute count,
are averaged across files, which is useful when you want to weight each
replicate evenly without regards to differing depth.}

\item{distancePower}{Length-one numeric vector determing to what power
Euclidean distance values are raised before fitting. Higher powers tend
to increase high values and decrease low values, exaggerating the
variation in the data. Default of 4.}

\item{peakIntervalWidth}{Length-one numeric vector between \code{0} and
\code{1} specifying desired width of linkage region(s). The default value
of \code{0.95}, for example, yields peak regions defined as including the
top 95\% of SNPs in the peak region, as determined by the peak
resampling distribution.}

\item{loessOptResolution}{Length-one numeric vector between \code{0} and
\code{1} specifying
desired resolution for Loess fit optimization. The default of \code{0.001},
for example, indicates that the span ultimately chosen will perform better
than its neighbor values at \code{+-0.001}.}

\item{loessOptCutFactor}{Length-one numeric vector between \code{0} and
\code{1} specifying how aggressively the Loess
optimization algorithm proceeds. For example, with a default of \code{0.1}
different spans at intervals of \code{0.001} would be evaluated after
intervals of \code{0.01}.}
}
\value{
A \code{MmapprParam} object.
}
\description{
MmapprParam Constructor
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    mmappr_param <- mmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                                mutFiles = MMAPPR2data::exampleMutBam(),
                                refFasta = MMAPPR2data::goldenFasta(),
                                gtf = MMAPPR2data::gtf(),
                                outputFolder = tempOutputFolder())
}

}
