% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_functions.R
\name{outputMmapprData}
\alias{outputMmapprData}
\title{Title}
\usage{
outputMmapprData(mmapprData, plotAicc = FALSE)
}
\arguments{
\item{mmapprData}{description}

\item{plotAicc}{description}
}
\description{
Title
}
