% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outputMmapprData}
\alias{outputMmapprData}
\title{Generate plots and tables from MMAPPR2 data}
\usage{
outputMmapprData(mmapprData)
}
\arguments{
\item{mmapprData}{The \linkS4class{MmapprData} object to be output}
}
\value{
A \linkS4class{MmapprData} object after writing output files
  to the folder specified in the \code{outputFolder} slot of the
  \code{link{MmapprParam}} used.
}
\description{
Generate plots and tables from MMAPPR2 data
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE) &
        Sys.which('vep') != '') {
    slc24a5genome <- gmapR::GmapGenome(MMAPPR2data::goldenFasta(),
                                       name = 'slc24a5',
                                       create = TRUE)
    # Specify parameters:
    mmapprParam <- MmapprParam(refGenome = slc24a5genome,
                               wtFiles = MMAPPR2data::exampleWTbam(),
                               mutFiles = MMAPPR2data::exampleMutBam(),
                               species = "danio_rerio",
                               outputFolder = tempdir())
}
\dontrun{
md <- new('MmapprData', param = mmapprParam)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)
postPrePeakMD <- prePeak(postLoessMD)
postPeakRefMD <- peakRefinement(postPrePeakMD)
postCandidatesMD <- generateCandidates(postPeakRefMD)

outputMmapprData(postCandidatesMD)
}
}
