% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{outputMmapprData}
\alias{outputMmapprData}
\title{Generate plots and tables from MMAPPR2 data}
\arguments{
\item{mmapprData}{The \linkS4class{MmapprData} object to be output}
}
\value{
A \linkS4class{MmapprData} object after writing output files
  to the folder specified in the \code{outputFolder} slot of the
  \code{link{MmapprParam}} used.
}
\description{
Generate plots and tables from MMAPPR2 data
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    mmappr_param <- mmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                                mutFiles = MMAPPR2data::exampleMutBam(),
                                refFasta = MMAPPR2data::goldenFasta(),
                                gtf = MMAPPR2data::gtf(),
                                outputFolder = tempOutputFolder())
}

\dontrun{
md <- mmapprData(mmappr_param)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)
postPrePeakMD <- prePeak(postLoessMD)
postPeakRefMD <- peakRefinement(postPrePeakMD)
postCandidatesMD <- generateCandidates(postPeakRefMD)

outputMmapprData(postCandidatesMD)
}

}
