% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peakRefinement}
\alias{peakRefinement}
\title{Characterize Euclidean distance peaks using resampling simulation}
\usage{
peakRefinement(mmapprData)
}
\arguments{
\item{mmapprData}{}
}
\value{
A \linkS4class{MmapprData} object with the \code{peaks}
  slot filled and populated.
}
\description{
Characterize Euclidean distance peaks using resampling simulation
}
