% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peakRefinement}
\alias{peakRefinement}
\title{Characterize Euclidean distance peaks using resampling simulation}
\usage{
peakRefinement(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{peaks}
  slot filled and populated.
}
\description{
Follows the \code{\link{prePeak}} step and precedes
\code{\link{generateCandidates}}.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE) &
        Sys.which('vep') != '') {
    slc24a5genome <- gmapR::GmapGenome(MMAPPR2data::goldenFasta(),
                                       name = 'slc24a5',
                                       create = TRUE)
    # Specify parameters:
    mmapprParam <- MmapprParam(refGenome = slc24a5genome,
                               wtFiles = MMAPPR2data::exampleWTbam(),
                               mutFiles = MMAPPR2data::exampleMutBam(),
                               species = "danio_rerio",
                               outputFolder = tempdir())
}
\dontrun{
md <- new('MmapprData', param = mmapprParam)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)
postPrePeakMD <- prePeak(postLoessMD)

postPeakRefMD <- peakRefinement(postPrePeakMD)
}
}
