% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peakRefinement}
\alias{peakRefinement}
\title{Characterize Euclidean distance peaks using resampling simulation}
\usage{
peakRefinement(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{peaks}
  slot filled and populated.
}
\description{
Follows the \code{\link{prePeak}} step and precedes
\code{\link{generateCandidates}}.
}
\examples{
\dontrun{
md <- peakRefinement(md)
}
}
