% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peakRefinement}
\alias{peakRefinement}
\title{Characterize Euclidean distance peaks using resampling simulation}
\usage{
peakRefinement(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{peaks}
  slot filled and populated.
}
\description{
Follows the \code{\link{prePeak}} step and precedes
\code{\link{generateCandidates}}.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly=TRUE)
        & all(Sys.which(c("samtools", "vep")) != "")) {
    mmappr_param <- MmapprParam(refFasta = MMAPPR2data::goldenFasta(),
                               wtFiles = MMAPPR2data::exampleWTbam(),
                               mutFiles = MMAPPR2data::exampleMutBam(),
                               species = "danio_rerio",
                               outputFolder = tempOutputFolder())
}
\dontrun{
md <- new('MmapprData', param = mmappr_param)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)
postPrePeakMD <- prePeak(postLoessMD)

postPeakRefMD <- peakRefinement(postPrePeakMD)
}
}
