% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peakRefinement}
\alias{peakRefinement}
\title{Characterize Euclidean distance peaks using resampling simulation}
\usage{
Follows the \code{\link{prePeak}} step and precedes
\code{\link{generateCandidates}}.
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \code{\linkS4class{MmapprData}} object with the \code{peaks}
  slot filled and populated.
}
\description{
Characterize Euclidean distance peaks using resampling simulation
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    mmappr_param <- mmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                                mutFiles = MMAPPR2data::exampleMutBam(),
                                refFasta = MMAPPR2data::goldenFasta(),
                                gtf = MMAPPR2data::gtf(),
                                outputFolder = tempOutputFolder())
}

\dontrun{
md <- mmapprData(mmappr_param)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)
postPrePeakMD <- prePeak(postLoessMD)

postPeakRefMD <- peakRefinement(postPrePeakMD)
}

}
