% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{prePeak}
\alias{prePeak}
\title{Identify chromosomes containing peaks}
\usage{
prePeak(mmapprData)
}
\arguments{
\item{mmapprData}{}
}
\value{
A \linkS4class{MmapprData} object with the \code{peaks}
  slot filled.
}
\description{
Identify chromosomes containing peaks
}
