% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{prePeak}
\alias{prePeak}
\title{Identify chromosomes containing peaks}
\usage{
prePeak(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \linkS4class{MmapprData} object with the \code{peaks}
  slot initalized.
}
\description{
Follows the \code{\link{loessFit}} step and precedes
\code{\link{peakRefinement}}.
}
\examples{
\dontrun{
md <- prePeak(md)
}
}
