% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{prePeak}
\alias{prePeak}
\title{Identify chromosomes containing peaks}
\usage{
prePeak(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \linkS4class{MmapprData} object with the \code{peaks}
  slot initalized.
}
\description{
Follows the \code{\link{loessFit}} step and precedes
\code{\link{peakRefinement}}.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    ## Ignore these lines:
    MMAPPR2:::.insertFakeVEPintoPath()
    genDir <- gmapR::GmapGenomeDirectory('DEFAULT', create=TRUE)

    # Specify parameters:
    mmapprParam <- MmapprParam(refGenome = gmapR::GmapGenome("GRCz11", genDir),
                               wtFiles = MMAPPR2data::zy13wtBam(),
                               mutFiles = MMAPPR2data::zy13mutBam(),
                               species = "danio_rerio")
}
\dontrun{
md <- new('MmapprData', param = mmapprParam)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)

postPrePeakMD <- prePeak(postLoessMD)
}
}
