% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{prePeak}
\alias{prePeak}
\title{Identify chromosomes containing peaks}
\usage{
prePeak(mmapprData)
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \linkS4class{MmapprData} object with the \code{peaks}
  slot initalized.
}
\description{
Follows the \code{\link{loessFit}} step and precedes
\code{\link{peakRefinement}}.
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE) &
        Sys.which('vep') != '') {
    genDir <- gmapR::GmapGenomeDirectory(tempdir(), create=TRUE)

    # Specify parameters:
    mmapprParam <- MmapprParam(refGenome = gmapR::GmapGenome("GRCz11", genDir),
                               wtFiles = MMAPPR2data::exampleWTbam(),
                               mutFiles = MMAPPR2data::exampleMutBam(),
                               species = "danio_rerio",
                               outputFolder = tempdir())
}
\dontrun{
md <- new('MmapprData', param = mmapprParam)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)

postPrePeakMD <- prePeak(postLoessMD)
}
}
