% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{prePeak}
\alias{prePeak}
\title{Identify chromosomes containing peaks}
\usage{
Follows the \code{\link{loessFit}} step and precedes
\code{\link{peakRefinement}}.
}
\arguments{
\item{mmapprData}{The \code{\linkS4class{MmapprData}} object to be analyzed.}
}
\value{
A \linkS4class{MmapprData} object with the \code{peaks}
  slot initalized.
}
\description{
Identify chromosomes containing peaks
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    mmappr_param <- mmapprParam(wtFiles = MMAPPR2data::exampleWTbam(),
                                mutFiles = MMAPPR2data::exampleMutBam(),
                                refFasta = MMAPPR2data::goldenFasta(),
                                gtf = MMAPPR2data::gtf(),
                                outputFolder = tempOutputFolder())
}

\dontrun{
md <- mmapprData(mmappr_param)
postCalcDistMD <- calculateDistance(md)
postLoessMD <- loessFit(postCalcDistMD)

postPrePeakMD <- prePeak(postLoessMD)
}

}
