% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tempOutputFolder}
\alias{tempOutputFolder}
\title{Generate temporary output folder}
\usage{
tempOutputFolder()
}
\value{
The path to the temporary directory
}
\description{
Conveniently creates a timestamp-named temporary directory
}
\examples{
mmapprParam <- MmapprParam(refGenome = slc24a5genome,
                           wtFiles = MMAPPR2data::exampleWTbam(),
                           mutFiles = MMAPPR2data::exampleMutBam(),
                           species = "danio_rerio",
                           outputFolder = tempOutputFolder())
}
