% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output.R
\name{tempOutputFolder}
\alias{tempOutputFolder}
\title{Generate temporary output folder}
\usage{
Conveniently creates a timestamp-named temporary directory
}
\value{
The path to the temporary directory
}
\description{
Generate temporary output folder
}
\examples{
if (requireNamespace('MMAPPR2data', quietly = TRUE)) {
    mmappr_param <- MmapprParam(refFasta = MMAPPR2data::goldenFasta(),
                                wtFiles = MMAPPR2data::exampleWTbam(),
                                mutFiles = MMAPPR2data::exampleMutBam(),
                                species = "danio_rerio",
                                outputFolder = tempOutputFolder())
}
}
